Fixes for types of props

+++ Models/Base/ProblemBaseModel.cs
@@ -9,7 +9,7 @@ namespace ContestSystem.Models.Base
     public class ProblemBaseModel : BaseProblem
     {
         public override long Id { get; set; }
-        public override int MemoryLimit { get; set; }
+        public override long MemoryLimit { get; set; }
         public override int TimeLimit { get; set; }
         public override ProblemType Type { get; set; }
         public string Name { get; set; }
